% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data1.R
\docType{data}
\name{data1}
\alias{data1}
\title{data1}
\format{
Tibble frame with columns
\describe{#colum name
\item{Company}{Two letter abbreviation.}
\item{Year}{Full name.}
\item{Sales}{Full name.}
}
}
\source{
\url{https://www.transtats.bts.gov/DL_SelectFields.asp?Table_ID=236}
}
\usage{
data1
}
\description{
Sales Volume of Korean Companies for three years
}
\examples{
  data1
}
\keyword{datasets}
