% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2.R
\docType{data}
\name{data2}
\alias{data2}
\title{data2}
\format{
Tibble frame with columns
\describe{#colum name
\item{Company}{Two letter abbreviation.}
\item{Year}{Full name.}
\item{Debt}{Full name.}
}
}
\source{
\url{https://www.transtats.bts.gov/DL_SelectFields.asp?Table_ID=236}
}
\usage{
data2
}
\description{
Debt Volume of Korean Companies for three years
}
\examples{
  data2
}
\keyword{datasets}
