% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invest-package.R
\docType{package}
\name{invest-package}
\alias{invest}
\alias{invest-package}
\title{invest: company inveset data file}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SeungMinK/invest}
  \item Report bugs at \url{https://github.com/SeungMinK/invest/issues}
}

}
\keyword{internal}
